requirements.txt
langchain==0.0.300
pandas==2.1.0
python-dotenv==1.0.0